eb-4953: delete salesrep cookie on logout or timeout
XPEDXSessionManageFilter no longer creates the salespro cookie, it only deletes it on a timeout. instead, the cookie is created as salespro performs 'login as customer' operation.
also removed unneeded ajax call to XPEDXSalesRepLoginAction. this was called via ajax from salespro login page attempting to delete cookies and invalidate the HttpSession. the cookie logic was not implemented correctly and had no effect and the invalidation of the session is not desired functionality.